fast_api_whisper/pyproject.toml
[project]
name = "fast-api-whisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "whisper>=1.1.10",
    "yt-dlp>=2025.4.30",
]
